refactor: rename NMI gateway and provider
