+++ storefronts/features/checkout/gateways/index.js
@@ -1,7 +1,7 @@
 import * as stripe from './stripe.js';
 import * as authorizeNet from './authorizeNet.js';
 import * as paypal from './paypal.js';
-import * as nmi from './nmi.js';
+import * as nmi from './nmiGateway.js';
 import * as segpay from './segpay.js';
 
 export default {
