+++ storefronts/tests/providers/provider-nmi-mount.test.ts
@@ -73,7 +73,7 @@ beforeEach(async () => {
 
   window.SMOOTHR_CONFIG = { storeId: 'store-1', active_payment_gateway: 'nmi' } as any;
 
-  const mod = await import('../../features/checkout/gateways/nmi.js');
+  const mod = await import('../../features/checkout/gateways/nmiGateway.js');
   mountNMI = mod.mountNMI;
   ready = mod.ready;
 });
